retreat dialog
